+++ src/comments.js
@@ -58,7 +58,7 @@ function getSortedChildNodes(node, text, resultArray) {
 
   if (!resultArray) {
     Object.defineProperty(node, childNodesCacheKey, {
-      value: resultArray = [],
+      value: (resultArray = []),
       enumerable: false
     });
   }
@@ -143,10 +143,23 @@ function attach(comments, ast, text, options) {
       // If a comment exists on its own line, prefer a leading comment.
       // We also need to check if it's the first line of the file.
       if (
+        handleLastFunctionArgComments(
+          precedingNode,
+          enclosingNode,
+          followingNode,
+          comment
+        ) ||
         handleMemberExpressionComments(enclosingNode, followingNode, comment) ||
         handleIfStatementComments(enclosingNode, followingNode, comment) ||
         handleTryStatementComments(enclosingNode, followingNode, comment) ||
-        handleClassComments(enclosingNode, comment)
+        handleImportSpecifierComments(enclosingNode, comment) ||
+        handleClassComments(enclosingNode, comment) ||
+        handleUnionTypeComments(
+          precedingNode,
+          enclosingNode,
+          followingNode,
+          comment
+        )
       ) {
         // We're good
       } else if (followingNode) {
@@ -169,8 +182,11 @@ function attach(comments, ast, text, options) {
           comment,
           text
         ) ||
+        handleImportSpecifierComments(enclosingNode, comment) ||
         handleTemplateLiteralComments(enclosingNode, comment) ||
-        handleClassComments(enclosingNode, comment)
+        handleCallExpressionComments(precedingNode, enclosingNode, comment) ||
+        handleClassComments(enclosingNode, comment) ||
+        handlePropertyComments(enclosingNode, comment)
       ) {
         // We're good
       } else if (precedingNode) {
@@ -457,11 +473,101 @@ function handleFunctionDeclarationComments(enclosingNode, comment) {
   return false;
 }
 
+function handleLastFunctionArgComments(
+  precedingNode,
+  enclosingNode,
+  followingNode,
+  comment
+) {
+  // Type definitions functions
+  if (
+    precedingNode &&
+    precedingNode.type === "FunctionTypeParam" &&
+    enclosingNode &&
+    enclosingNode.type === "FunctionTypeAnnotation" &&
+    followingNode &&
+    followingNode.type !== "FunctionTypeParam"
+  ) {
+    addTrailingComment(precedingNode, comment);
+    return true;
+  }
+
+  // Real functions
+  if (
+    precedingNode &&
+    precedingNode.type === "Identifier" &&
+    enclosingNode &&
+    (enclosingNode.type === "ArrowFunctionExpression" ||
+      enclosingNode.type === "FunctionExpression") &&
+    followingNode &&
+    followingNode.type !== "Identifier"
+  ) {
+    addTrailingComment(precedingNode, comment);
+    return true;
+  }
+  return false;
+}
+
 function handleClassComments(enclosingNode, comment) {
   if (
     enclosingNode &&
     (enclosingNode.type === "ClassDeclaration" ||
       enclosingNode.type === "ClassExpression")
+  ) {
+    addLeadingComment(enclosingNode, comment);
+    return true;
+  }
+  return false;
+}
+
+function handleImportSpecifierComments(enclosingNode, comment) {
+  if (enclosingNode && enclosingNode.type === "ImportSpecifier") {
+    addLeadingComment(enclosingNode, comment);
+    return true;
+  }
+  return false;
+}
+
+function handleCallExpressionComments(precedingNode, enclosingNode, comment) {
+  if (
+    enclosingNode &&
+    enclosingNode.type === "CallExpression" &&
+    precedingNode &&
+    enclosingNode.callee === precedingNode &&
+    enclosingNode.arguments.length > 0
+  ) {
+    addLeadingComment(enclosingNode.arguments[0], comment);
+    return true;
+  }
+  return false;
+}
+
+function handleUnionTypeComments(
+  precedingNode,
+  enclosingNode,
+  followingNode,
+  comment
+) {
+  if (
+    enclosingNode &&
+    enclosingNode.type === "UnionTypeAnnotation" &&
+    precedingNode &&
+    precedingNode.type === "ObjectTypeAnnotation" &&
+    followingNode &&
+    followingNode.type === "ObjectTypeAnnotation"
+  ) {
+    addTrailingComment(precedingNode, comment);
+    return true;
+  }
+  return false;
+}
+
+function handlePropertyComments(enclosingNode, comment) {
+  if (
+    enclosingNode && (
+      enclosingNode.type === "Property" ||
+      enclosingNode.type === "ObjectProperty"
+    )
   ) {
     addLeadingComment(enclosingNode, comment);
     return true;