+++ tests/import/comments.js
@@ -1,0 +1,22 @@
+import { a //comment1
+//comment2
+//comment3
+as b} from "";
+
+import {
+  a as //comment1
+  //comment2
+  //comment3
+  b
+} from "";
+
+import {
+  a as //comment2 //comment1
+  //comment3
+  b
+} from "";
+
+import {
+  a as //comment3 //comment2 //comment1
+  b
+} from "";