+++ tests/require/require.js
@@ -1,0 +1,4 @@
+const { one, two, thee, four, five, six, seven, eight, nine, ten } = require('./my-utils');
+const { one, two, thee, four, five, six, seven, eight, nine, ten, eleven } = require('./my-utils');
+
+const MyReallyExtrememlyLongModuleName = require('MyReallyExtrememlyLongModuleName');